data_generator/structured_data/src/requirements.txt
pandas
random2
xlsxwriter
pillow